// server/vance-addon-brain-journal/client/src/generated/journal/JournalView.ts
/*
 Auto-generated by generate-java-to-ts-maven-plugin. Do not edit.
 Source: de.mhus.vance.addon.brain.journal.JournalView
*/

import { JournalEntryView } from './JournalEntryView';
import { JournalStatsView } from './JournalStatsView';

export interface JournalView {
  folder: string;
  title?: string;
  description?: string;
  entriesDir: string;
  moodPresets: string[];
  stats: JournalStatsView;
  recent: JournalEntryView[];
}