// server/vance-addon-brain-journal/client/src/generated/journal/JournalRebuildResponse.ts
/*
 Auto-generated by generate-java-to-ts-maven-plugin. Do not edit.
 Source: de.mhus.vance.addon.brain.journal.JournalRebuildResponse
*/

export interface JournalRebuildResponse {
  folder: string;
  indexPath?: string;
  statsPath?: string;
  entryCount: number;
  currentStreak: number;
  longestStreak: number;
}